refactor(dto): InputDTO를 dto 패키지에 분리
InputDTO를 dto 패키지에 분리 후, InputDTO를 사용하는 클래스의 import 문을 변경한다.
