+++ src/main/java/racingcar/dto/InputDTO.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.dto;
 
 import java.util.List;
 import racingcar.model.RacingCar;
