+++ src/test/java/racingcar/InputDTOTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import racingcar.dto.InputDTO;
 import racingcar.model.RacingCar;
 
 class InputDTOTest {
